+++ src/Components/Validator/CaseStyle/Exception/CaseStyleNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PHPUnuhi\Components\Validator\CaseStyle\Exception;
+
+class CaseStyleNotFoundException extends \Exception
+{
+
+}